+++ cai_causal_graph/time_series_causal_graph.py
@@ -282,6 +282,28 @@ class TimeSeriesCausalGraph(CausalGraph):
 
         return minimal_cg
 
+    def _get_time_topological_order(self, ordered_nodes: List[str]) -> List[str]:
+        """Return the time topological order given the standard topological order. Only for internal use."""
+        min_time_lag = max([node.time_lag for node in self.get_nodes()])
+
+        # extract only nodes at min_time_lag
+        cont_ordered_nodes = [
+            node for node in ordered_nodes if self.get_node(node).time_lag == min_time_lag  # type: ignore
+        ]
+        # get the remaining nodes
+        lagged_nodes = [n for n in ordered_nodes if self.get_node(n).time_lag != min_time_lag]  # type: ignore
+
+        # sort lagged nodes based on the order they appear in the non-lagged list and by their lag values
+        lagged_nodes = sorted(
+            lagged_nodes,
+            key=lambda x: (cont_ordered_nodes.index(self.get_node(x).variable_name), self.get_node(x).time_lag),  # type: ignore
+        )
+
+        # merge the lists together
+        ordered_nodes = lagged_nodes + cont_ordered_nodes
+
+        return ordered_nodes
+
     def get_topological_order(self, return_all: bool = False) -> Union[List[str], List[List[str]]]:
         """
         Return the topological order of the graph that is ordered in time.
@@ -291,15 +313,16 @@ class TimeSeriesCausalGraph(CausalGraph):
         """
         ordered_nodes = super().get_topological_order(return_all=return_all)
         if return_all:
-            # sort by time lag
-            ordered_nodes = [
-                sorted(vord, key=lambda x: self.get_node(x[0]).time_lag) for vord in ordered_nodes  # type: ignore
-            ]
+            top_order_list: List[List[str]] = []
+            for current_top_order in ordered_nodes:
+                assert isinstance(current_top_order, list)   # for linting
+                new_order = self._get_time_topological_order(current_top_order)
+                if new_order not in top_order_list:
+                    top_order_list.append(new_order)
         else:
-            # sort by time lag
-            ordered_nodes = sorted(ordered_nodes, key=lambda x: self.get_node(x).time_lag)  # type: ignore
+            top_order_list = self._get_time_topological_order(ordered_nodes)   # type: ignore
 
-        return ordered_nodes
+        return top_order_list
 
     def is_minimal_graph(self) -> bool:
         """